refactor profiler
