+++ localstack/utils/kinesis/java/cloud/localstack/KinesisStarter.java
@@ -3,6 +3,8 @@ package cloud.localstack;
 import java.util.Properties;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
+import java.util.logging.Logger;
+import java.util.logging.Level;
 
 import com.amazonaws.services.kinesis.clientlibrary.lib.worker.KinesisClientLibConfiguration;
 import com.amazonaws.services.kinesis.multilang.MultiLangDaemon;
