+++ localstack/utils/kinesis/kinesis_connector.py
@@ -5,6 +5,7 @@ import re
 import logging
 import tempfile
 import threading
+import subprocess32 as subprocess
 from six.moves import queue as Queue
 from six.moves.urllib.parse import urlparse
 from amazon_kclpy import kcl
@@ -14,7 +15,7 @@ from localstack.constants import (
     LOCALSTACK_VENV_FOLDER, LOCALSTACK_ROOT_FOLDER, DEFAULT_REGION)
 from localstack.utils.aws import aws_stack
 from localstack.utils.common import (
-    run, TMP_THREADS, TMP_FILES, save_file, now, retry, short_uid,
+    run, TMP_THREADS, TMP_FILES, save_file, now, retry, short_uid, to_str,
     chmod_r, rm_rf, ShellCommandThread, FuncThread)
 from localstack.utils.kinesis import kclipy_helper
 from localstack.utils.aws.aws_models import KinesisStream
@@ -130,7 +131,6 @@ class KinesisProcessorThread(ShellCommandThread):
 
 
 class OutputReaderThread(FuncThread):
-    OUTPUT_LINES_COUNT = 0
 
     def __init__(self, params):
         FuncThread.__init__(self, self.start_reading, params)
@@ -183,11 +183,6 @@ class OutputReaderThread(FuncThread):
 
     def start_reading(self, params):
         for line in self._tail(params['file']):
-            OutputReaderThread.OUTPUT_LINES_COUNT += 1
-            if OutputReaderThread.OUTPUT_LINES_COUNT % 100 == 0:
-                print('Received %s log lines from Kinesis clients' %
-                      OutputReaderThread.OUTPUT_LINES_COUNT)
-
             # notify subscribers
             self.notify_subscribers(line)
             if self.log_level > 0:
@@ -205,6 +200,16 @@ class OutputReaderThread(FuncThread):
                     self.buffer = []
 
     def _tail(self, file):
+        p = subprocess.Popen(['tail', '-f', file], stdout=subprocess.PIPE)
+        while True:
+            line = p.stdout.readline()
+            if not line:
+                break
+            line = to_str(line)
+            yield line.replace('\n', '')
+
+    def _tail_native(self, file):
+        # deprecated
         with open(file) as f:
             while self.running:
                 line = f.readline()
@@ -346,11 +351,11 @@ def generate_processor_script(events_file, log_file=None):
         log_file = 'None'
     content = """#!/usr/bin/env python
 import os, sys, glob, json, socket, time, logging, tempfile
-import subprocess32 as subprocess
 logging.basicConfig(level=logging.INFO)
 for path in glob.glob('%s/lib/python*/site-packages'):
     sys.path.insert(0, path)
 sys.path.insert(0, '%s')
+import subprocess32 as subprocess
 from localstack.config import DEFAULT_ENCODING
 from localstack.utils.kinesis import kinesis_connector
 from localstack.utils.common import timestamp
